attempt to streamline Steemchains
